[FE] 일기 상세 조회 및 수정 API 연동

+++ frontend/src/pages/Detail.tsx
@@ -1,4 +1,8 @@
 import { useState } from 'react';
+import { useParams, useNavigate } from 'react-router-dom';
+import { useQuery } from '@tanstack/react-query';
+
+import referDiary from '@api/ReferDiary';
 
 import NavBar from '@components/Common/NavBar';
 import Button from '@components/Common/Button';
@@ -6,28 +10,51 @@ import Modal from '@components/Common/Modal';
 import DiaryContent from '@components/Detail/DiaryContent';
 import Alert from '@components/Common/Alert';
 
+import { PAGE_URL } from '@util/constants';
+
 const Detail = () => {
+  const navigate = useNavigate();
   const [showModal, setShowModal] = useState(false);
   const toggleShowModal = () => setShowModal((prev) => !prev);
 
-  const diaryData = {
-    createdAt: '2023-11-13T13:50:17.106Z',
-    profileImage: '',
-    authorName: '종현',
-    nickname: '단디',
-    title: '시고르자브종',
-    content: `일기내용입니다.\n일기내용입니다.\n다다다다다다다다다다다다다다다다다다다다\n다다다다다다다다다다다다다다다다다다다다\n다다다다다다다다다다다다다다다다다다다다\n\n`,
-    keywords: ['키워드1', '키워드2', '키워드3', '키워드4'],
-    reactionCount: 10,
-  };
-  const content = <DiaryContent {...diaryData} />;
+  const params = useParams();
+  const diaryId = Number(params.diaryId);
+  const { data, isLoading, isError } = useQuery({
+    queryKey: ['diary', diaryId],
+    queryFn: () => referDiary(diaryId),
+  });
+
+  if (isLoading) {
+    return <p>Loading...</p>;
+  }
+
+  if (isError) {
+    return <p>Error fetching data</p>;
+  }
+
+  const content = <DiaryContent {...data} />;
 
   return (
     <div className="flex flex-col items-center">
       <NavBar />
       <div className="flex w-2/3 flex-col gap-2">
         <div className="flex justify-end gap-2">
-          <Button text="수정" type="normal" onClick={() => console.log('수정하러 가기')} />
+          <Button
+            text="수정"
+            type="normal"
+            onClick={() =>
+              navigate(PAGE_URL.EDIT, {
+                state: {
+                  diaryId: diaryId,
+                  title: data.title,
+                  content: data.content,
+                  emotion: data.emotion,
+                  tags: data.tags,
+                  status: data.status,
+                },
+              })
+            }
+          />
           <Button text="삭제" type="delete" onClick={toggleShowModal} />
           <Modal showModal={showModal} closeModal={toggleShowModal}>
             <Alert
